% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{align_samples}
\alias{align_samples}
\title{Align expression and clinical samples}
\usage{
align_samples(expr, clinical)
}
\arguments{
\item{expr}{Expression matrix with samples in columns.}

\item{clinical}{Clinical data frame with samples in row names.}
}
\value{
A list with aligned \code{expr} and \code{clinical}.
}
\description{
Align expression and clinical samples
}
