% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_expression_matrix}
\alias{filter_expression_matrix}
\title{Filter expression matrix by non-zero expression}
\usage{
filter_expression_matrix(expr, min_nonzero_prop = 0.5)
}
\arguments{
\item{expr}{Expression data frame or matrix with genes in rows and samples in columns.}

\item{min_nonzero_prop}{Minimum proportion of samples with expression greater than zero.}
}
\value{
A numeric matrix.
}
\description{
Filter expression matrix by non-zero expression
}
