% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunePair-package.R
\docType{package}
\name{immunePair-package}
\alias{immunePair}
\alias{immunePair-package}
\title{immunePair}
\description{
Tools for pairwise gene marker analysis in immune response studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/twl-00/immunePair}
  \item Report bugs at \url{https://github.com/twl-00/immunePair/issues}
}

}
\author{
\strong{Maintainer}: twl-00 \email{tanwenlu2000@163.com}

}
\keyword{internal}
