% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairMarker-package.R
\docType{package}
\name{PairMarker-package}
\alias{PairMarker}
\alias{PairMarker-package}
\title{PairMarker}
\description{
Tools for pairwise gene marker discovery using binary phenotype labels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/twl-00/PairMarker}
  \item Report bugs at \url{https://github.com/twl-00/PairMarker/issues}
}

}
\author{
\strong{Maintainer}: twl-00 \email{tanwenlu2000@163.com}

}
\keyword{internal}
