% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_clinical_data}
\alias{read_clinical_data}
\title{Read clinical data}
\usage{
read_clinical_data(file, sep = "\\t")
}
\arguments{
\item{file}{Path to a tab-delimited clinical file.}

\item{sep}{Field separator.}
}
\value{
A data frame with samples as row names.
}
\description{
Read clinical data
}
