% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Read expression data}
\usage{
read_expression_data(file, gene_col = NULL, sep = "\\t")
}
\arguments{
\item{file}{Path to a tab-delimited expression file.}

\item{gene_col}{Name of the gene column. If \code{NULL}, the first column is used
when row names are not already meaningful.}

\item{sep}{Field separator.}
}
\value{
A data frame with genes as row names and samples as columns.
}
\description{
Read expression data
}
