% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_analysis.R
\name{run_pairwise_chisq}
\alias{run_pairwise_chisq}
\title{Run C++ pairwise chi-square screening}
\usage{
run_pairwise_chisq(
  expr,
  resp,
  min_prop = 0.05,
  max_prop = 0.95,
  chisq_cutoff = 0.01,
  diff_cutoff = 0.25,
  min_valid_prop = 0.5
)
}
\arguments{
\item{expr}{Numeric expression matrix with genes in rows and samples in columns.}

\item{resp}{Integer response vector with 1 for response and 0 for non-response.}

\item{min_prop}{Minimum proportion of pairwise comparison value 1.}

\item{max_prop}{Maximum proportion of pairwise comparison value 1.}

\item{chisq_cutoff}{Chi-square p-value cutoff.}

\item{diff_cutoff}{Minimum expression difference for assigning pairwise order.}

\item{min_valid_prop}{Minimum proportion of valid samples per gene pair.}
}
\value{
A data frame with screened gene pairs and chi-square p-values.
}
\description{
Run C++ pairwise chi-square screening
}
